fix-a-warn on Solaris 9 (http://sphinxsearch.com/forum/view.html?id=102)
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@463 406a0c4d-033a-0410-8de8-e80135713968

+++ src/searchd.cpp
@@ -253,7 +253,7 @@ void sphLog ( ESphLogLevel eLevel, const char * sFmt, va_list ap )
 
 	char sBuf [ 1024 ];
 	if ( !isatty ( g_iLogFile ) )
-		snprintf ( sBuf, sizeof(sBuf)-1, "[%s] [%5d] %s", sTimeBuf, getpid(), sBanner );
+		snprintf ( sBuf, sizeof(sBuf)-1, "[%s] [%5d] %s", sTimeBuf, (int)getpid(), sBanner );
 	else
 		strcpy ( sBuf, sBanner );
 	int iLen = strlen(sBuf);
